+++ modules/Accounting/Expenses.php
@@ -34,12 +34,10 @@ if ( ! empty( $_REQUEST['values'] )
 		elseif ( $columns['AMOUNT'] !== ''
 			&& $columns['PAYMENT_DATE'] )
 		{
-			$id = DBSeqNextID( 'accounting_payments_id_seq' );
-
 			$sql = "INSERT INTO ACCOUNTING_PAYMENTS ";
 
-			$fields = 'ID,SYEAR,SCHOOL_ID,';
-			$values = "'" . $id . "','" . UserSyear() . "','" . UserSchool() . "',";
+			$fields = 'SYEAR,SCHOOL_ID,';
+			$values = "'" . UserSyear() . "','" . UserSchool() . "',";
 
 			if ( isset( $_FILES['FILE_ATTACHED'] ) )
 			{