+++ modules/Food_Service/Users/ServeMenus.php
@@ -23,27 +23,47 @@ if ( $_REQUEST['modfunc'] === 'submit' )
 	elseif ( $_REQUEST['submit']['save']
 		&& ! empty( $_SESSION['FSA_sale'] ) )
 	{
-		$items_RET = DBGet( "SELECT DESCRIPTION,SHORT_NAME,PRICE_STAFF FROM FOOD_SERVICE_ITEMS WHERE SCHOOL_ID='" . UserSchool() . "'", [], [ 'SHORT_NAME' ] );
+		$fields = 'STAFF_ID,SYEAR,SCHOOL_ID,BALANCE,TIMESTAMP,SHORT_NAME,DESCRIPTION,SELLER_ID';
 
-		// get next transaction id
-		$id = DBSeqNextID( 'food_service_staff_transactions_transaction_id_seq' );
+		$values = "'" . UserStaffID() . "','" . UserSyear() . "','" . UserSchool() .
+			"',(SELECT BALANCE
+			FROM FOOD_SERVICE_STAFF_ACCOUNTS
+			WHERE STAFF_ID='" . UserStaffID() . "'),CURRENT_TIMESTAMP,'" .
+			$menus_RET[$_REQUEST['menu_id']][1]['TITLE'] . "','" .
+			$menus_RET[$_REQUEST['menu_id']][1]['TITLE'] . ' - ' . DBDate() . "','" . User( 'STAFF_ID' ) . "'";
+
+		$sql = "INSERT INTO FOOD_SERVICE_STAFF_TRANSACTIONS (" . $fields . ") values (" . $values . ")";
+
+		DBQuery( $sql );
+
+		$transaction_id = DBLastInsertID();
+
+		$items_RET = DBGet( "SELECT DESCRIPTION,SHORT_NAME,PRICE_STAFF
+			FROM FOOD_SERVICE_ITEMS
+			WHERE SCHOOL_ID='" . UserSchool() . "'", [], [ 'SHORT_NAME' ] );
 
 		$item_id = 0;
 
 		foreach ( (array) $_SESSION['FSA_sale'] as $item_sn )
 		{
 			$price = $items_RET[$item_sn][1]['PRICE_STAFF'];
+
 			$fields = 'ITEM_ID,TRANSACTION_ID,AMOUNT,SHORT_NAME,DESCRIPTION';
-			$values = "'" . $item_id++ . "','" . $id . "','-" . $price . "','" . $items_RET[$item_sn][1]['SHORT_NAME'] . "','" . $items_RET[$item_sn][1]['DESCRIPTION'] . "'";
+
+			$values = "'" . $item_id++ . "','" . $transaction_id . "','-" . $price . "','" . $items_RET[$item_sn][1]['SHORT_NAME'] . "','" . $items_RET[$item_sn][1]['DESCRIPTION'] . "'";
+
 			$sql = "INSERT INTO FOOD_SERVICE_STAFF_TRANSACTION_ITEMS (" . $fields . ") values (" . $values . ")";
+
 			DBQuery( $sql );
 		}
 
-		$sql1 = "UPDATE FOOD_SERVICE_STAFF_ACCOUNTS SET TRANSACTION_ID='" . (int) $id . "',BALANCE=BALANCE+(SELECT sum(AMOUNT) FROM FOOD_SERVICE_STAFF_TRANSACTION_ITEMS WHERE TRANSACTION_ID='" . (int) $id . "') WHERE STAFF_ID='" . UserStaffID() . "'";
-		$fields = 'TRANSACTION_ID,STAFF_ID,SYEAR,SCHOOL_ID,BALANCE,TIMESTAMP,SHORT_NAME,DESCRIPTION,SELLER_ID';
-		$values = "'" . $id . "','" . UserStaffID() . "','" . UserSyear() . "','" . UserSchool() . "',(SELECT BALANCE FROM FOOD_SERVICE_STAFF_ACCOUNTS WHERE STAFF_ID='" . UserStaffID() . "'),CURRENT_TIMESTAMP,'" . $menus_RET[$_REQUEST['menu_id']][1]['TITLE'] . "','" . $menus_RET[$_REQUEST['menu_id']][1]['TITLE'] . ' - ' . DBDate() . "','" . User( 'STAFF_ID' ) . "'";
-		$sql2 = 'INSERT INTO FOOD_SERVICE_STAFF_TRANSACTIONS (' . $fields . ') values (' . $values . ')';
-		DBQuery( 'BEGIN; ' . $sql1 . '; ' . $sql2 . '; COMMIT' );
+		$sql = "UPDATE FOOD_SERVICE_STAFF_ACCOUNTS
+			SET TRANSACTION_ID='" . (int) $transaction_id . "',BALANCE=BALANCE+(SELECT sum(AMOUNT)
+				FROM FOOD_SERVICE_STAFF_TRANSACTION_ITEMS
+				WHERE TRANSACTION_ID='" . (int) $transaction_id . "')
+			WHERE STAFF_ID='" . UserStaffID() . "'";
+
+		DBQuery( $sql );
 
 		unset( $_SESSION['FSA_sale'] );