+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 9.2.1
 - SQL rename character varying & character data types to varchar & char in rosariosis.sql
 - SQL replace use of STRPOS() with LIKE, compatible with MySQL in PortalPollNotes.fnc.php & Courses.php
 - SQL fix French & Spanish translation for Create Parent Users & Notifiy Parents email templates in rosariosis_fr.sql & rosariosis_es.sql
+- Use DBLastInsertID() instead of DBSeqNextID(), program wide
 
 Changes in 9.2
 --------------